refactor: move codex modules under src/codex
