+++ src/codex/codex-output-parser.ts
@@ -1,4 +1,4 @@
-import { type BlockedReason, type RunState } from "./types";
+import { type BlockedReason, type RunState } from "../types";
 
 const SUPPORTED_RUN_STATES: RunState[] = [
   "queued",
