+++ src/codex/index.ts
@@ -6,4 +6,4 @@ export {
   type LocalReviewRepairContext,
 } from "./codex-prompt";
 export { runCodexTurn } from "./codex-runner";
-export type { CodexTurnResult, IssueRunRecord, RunState, SupervisorConfig } from "./types";
+export type { CodexTurnResult, IssueRunRecord, RunState, SupervisorConfig } from "../types";
